move address to bech func into utils
